help/requirements.txt
Flask==2.3.0
flask-mysqldb==1.0.1
qrcode==7.4.2
Pillow==10.0.0
mysqlclient==2.2.0
